refactor: move input streaming to module
